improve partionning

+++ src/main/scala/settings/Settings.scala
@@ -11,5 +11,5 @@ class Settings(config: Config) {
     val rpath = config.getString("app.regressionPath")
     val densitiyInterspacEvaluation = config.getDouble("app.densitiyInterspacEvaluation")
     val defaultKernelBandWidth = config.getDouble("app.defaultKernelBandWidth")
-    val numberOfCores = config.getInt("app.numberOfCores")
+    val numberOfCores = config.getDouble("app.numberOfCores")
 }
